# api_server/app/src/auth/dto/sign_up_response.dto.ts
import { ApiProperty } from '@nestjs/swagger';

export class SignUpResponseDto {
  @ApiProperty({
    example: ['メールアドレスは必須です。', 'パスワードは必須です。'],
  })
  errors: Array<string>;
}